Join custom logic into bridge_commitments
